+++ src/utils/defaults/default_selectNames.js
@@ -1,2 +1,2 @@
 
-export const selectBookingPlattformPreName = "select_bookingPlattform_"
+export const selectBookingPlatformPreName = "select_bookingPlatform_"